feat(select): refactor Select components and add context for improved state management
- Separated existing select component into different files
- Add proper context and kept same style and functionality
- Update imports and exports
