+++ src/components/select/index.ts
@@ -1,1 +1,4 @@
-export { SelectRoot, SelectTrigger, SelectContent, SelectItem } from "./Select";
+export { SelectRoot } from "./primitives/SelectRoot";
+export { SelectTrigger } from "./primitives/SelectTrigger";
+export { SelectContent } from "./primitives/SelectContent";
+export { SelectItem } from "./primitives/SelectItem";
